Add a script to make checking templates easier.
It copies a bunch of code from test_entities_exist() in test_templates.py,
which is modified to instead fetch labels for entity IDs.

It expects one or more template names to be provided as command line arguments.

After fetching the labels, it repeats the loop (not ideal, I know) to replace
the IDs with the label + ID and then finally pretty-prints
the template object. I originally tried the (built-in?) pprint module
but didn't like the format of the output nor does it appear to support
colour, so I went with something else that I liked more.